requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.6
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
django-filter==1.1.0
djangorestframework==3.6.4
djangorestframework-gis==0.11.2
httpie==0.9.9
idna==2.6
Markdown==2.6.9
psycopg2==2.7.3.1
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
